pyproject.toml
[project]
name = "new-gym"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amqp==5.3.1",
    "asgiref==3.10.0",
    "billiard==4.2.2",
    "brotli==1.1.0",
    "celery==5.4.0",
    "cffi==1.17.1",
    "click==8.3.0",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "cron-descriptor==2.0.6",
    "cssselect2==0.7.0",
    "diff-match-patch==20241021",
    "django==5.0.6",
    "django-appconf==1.1.0",
    "django-auto-prefetching==0.2.12",
    "django-celery-beat==2.8.1",
    "django-cors-headers==4.4.0",
    "django-crispy-forms==2.2",
    "django-crontab==0.7.1",
    "django-dbbackup==5.0.1",
    "django-debug-toolbar==4.4.2",
    "django-extensions==3.2.3",
    "django-filter==24.2",
    "django-html==0.1.0",
    "django-htmx==1.18.0",
    "django-import-export==4.1.0",
    "django-js-asset==2.2.0",
    "django-maintenance-mode==0.21.1",
    "django-mptt==0.16.0",
    "django-rest-multiple-models==2.1.3",
    "django-select2==8.2.1",
    "django-simple-history==3.7.0",
    "django-tables2==2.7.0",
    "django-timezone-field==7.1",
    "django-widget-tweaks==1.5.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "fonttools==4.60.1",
    "kombu==5.6.0",
    "packaging==25.0",
    "pillow==10.3.0",
    "prompt-toolkit==3.0.52",
    "psycopg2==2.9.10",
    "pycparser==2.23",
    "pydyf==0.11.0",
    "pyjwt==2.10.1",
    "pyphen==0.15.0",
    "python-crontab==3.3.0",
    "python-dateutil==2.9.0.post0",
    "python-fsutil==0.15.0",
    "redis==5.2.0",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "tablib==3.5.0",
    "tinycss2==1.4.0",
    "tinyhtml5==2.0.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "vine==5.1.0",
    "wcwidth==0.2.14",
    "weasyprint==63.0",
    "webencodings==0.5.1",
    "whitenoise==6.7.0",
    "zopfli==0.2.3.post1",
]


[tool.basedpyright]
useLibraryCodeForTypes = true
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportOptionalCall = false
reportAttributeAccessIssue = false
reportOperatorIssue = false
reportCallIssue = false
reportArgumentType = false
reportAssignmentType = false
reportGeneralTypeIssues = false
reportOptionalMemberAccess = false
